Added Multiplayer Functionality
Can now play with multiple players locally and zombies will track the nearest player. This is done in FixedUpdate, maybe it should be done less frequently, not sure yet. This involved changing the Player and PlayerObject variables in the GlobalData class to lists and the enemies will track the closest one themselves
Also changed the wall texture and added more sprites to make it more complex.