Implemented mouse controls
Added fun firing capabilities for mouse and keyboard. Changed how the system worked slightly so there is now a bool value in MBPlayer that is true when the player is firing, false otherwise. Only shoot bullets when firing is true. This is kind of sloppy and may need changing in the future