Animations, Rifle, & Mouse Controls (#1)
* Added runtime animation loader

Items in shop now have the back-end in place to visually show what the items are. Only have the sprite done for the speed upgrade but will be able to plug in the new animator controllers once the sprites are created

* Added rifle to shop

Added the first new weapon, the rifle. The rifle has limited ammo, so when you run out of ammo you will be given your unlimited ammo pistol back.
Gun stats subject to change

* Added upgrade sprites/animations

Created animations for HealthUp and Heal1. Also changed the total animation time for each upgrade to 1 second instead of 2 seconds (it will switch between its two images every 0.5 seconds instead of every 1 second)

* Implemented mouse controls

Added gun firing capabilities for mouse and keyboard. Changed how the system worked slightly so there is now a bool value in MBPlayer that is true when the player is firing, false otherwise. Only shoot bullets when firing is true. This is kind of sloppy and may need changing in the future