Implemented permanent Gun modifiers, set up enums for guns and enemies, added XML comments, increased num enemies spawned per wave, rebalanced health/damage
Set up attributes for firing delay and damage in Player class so now when those attributes are being applied it will consider the gun values as well as the player values (set from upgrades).
Set up enums for guns and enemies to streamline the instantiation process. Now if you want to instantiate a new enemy or Gun you can call the constructor and provide the enum that you want. This will be how all instantiation is done to ensure that any changes in that constructor will be applied to any gun/enemy and I won't need to change things in multiple places.
Added a bunch of XML comments
Increased number of enemies spawned per wave to 5*waveNum
Greatly increased the scale of health and damage for enemies