Different Enemy Type Functionality & Fixed Enemy Speed
Can now add enemy types to the enum in Character.cs along with stats in the Character constructor that uses the Enemy enum to create a new enemy type. Currently we have a basic all around enemy and a tank enemy, more to come. When making an enemy constructor, there is a variable "type" in MBEnemy.cs which is a drop down of all enemy types that can be chosen from. All enemy prefabs need to be added to the array in the spawner script in the scene.